+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for the dividend timeline grouped by year
+- Added support for the investment timeline grouped by year
 - Set up the language localization for Français (`fr`)
 - Set up the language localization for Português (`pt`)